Add AMD GPU support, improve build system (#49)
* Build Linux full SYCL version such that Nvidia can run on SYCL

* define a macro for fetching dependencies

* Add macros to cmakelists for repeated blocks; fix conversion warnings

* Add commandline options for SYCL backend

* Add SYCL commandline build

* use static linking of onemath

* adjust launch script and building because onemkl doesn't work as static

* Build for AMD with ROCM backend

* Let the executable name for some of the linux builds be controlled by the -DEXECUTABLE_NAME commandline option for cmake

* split cmake files by OS, make build targets easier to set

* improve fallback conditions if MKL or OpenMP aren't found

* Fix linux cli targets

* Update readme

* default to intel if no SYCL backend specified

* fix windows build

* fix mac build

* update cluster script

+++ Source/LightwaveExplorerCore.cu
@@ -6557,7 +6557,7 @@ namespace hostFunctions{
 		case functionID("applyOptic"):
 			interpretParameters(cc, 1, iBlock, vBlock, parameters, defaultMask);
 			d.reset(sCPU);
-			if((*sCPU).optics.size() < static_cast<int>(parameters[0])+1){
+			if((*sCPU).optics.size() < static_cast<std::size_t>(parameters[0])+1){
 				throw std::runtime_error("applyOptic: optic is not loaded");
 			}
 			applyOptic(
@@ -6571,7 +6571,7 @@ namespace hostFunctions{
 		case functionID("applyOpticX"):
 			interpretParameters(cc, 1, iBlock, vBlock, parameters, defaultMask);
 			d.reset(sCPU);
-			if((*sCPU).optics.size() < static_cast<int>(parameters[0])+1){
+			if((*sCPU).optics.size() < static_cast<std::size_t>(parameters[0])+1){
 				throw std::runtime_error("applyOpticX: optic is not loaded");
 			}
 			applyOptic(
@@ -6585,7 +6585,7 @@ namespace hostFunctions{
 		case functionID("applyOpticY"):
 			interpretParameters(cc, 1, iBlock, vBlock, parameters, defaultMask);
 			d.reset(sCPU);
-			if((*sCPU).optics.size() < static_cast<int>(parameters[0])+1){
+			if((*sCPU).optics.size() < static_cast<std::size_t>(parameters[0])+1){
 				throw std::runtime_error("applyOpticY: optic is not loaded");
 			}
 			applyOptic(
